+++ app/src/main/java/woowacourse/movie/ui/home/adapter/TheaterAdapter.kt
@@ -1,6 +1,5 @@
 package woowacourse.movie.ui.home.adapter
 
-import android.content.Intent
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
@@ -8,13 +7,11 @@ import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
 import woowacourse.movie.databinding.ItemTheaterBinding
 import woowacourse.movie.model.movie.Theater
-import woowacourse.movie.ui.home.MovieHomeKey
-import woowacourse.movie.ui.reservation.MovieReservationActivity
 
 class TheaterAdapter(
     private val theaters: List<Theater>,
     private val movieContentId: Long,
-) : RecyclerView.Adapter<TheaterAdapter.TheaterViewHolder>() {
+) : RecyclerView.Adapter<TheaterViewHolder>() {
     override fun onCreateViewHolder(
         parent: ViewGroup,
         viewType: Int,
@@ -38,22 +35,4 @@ class TheaterAdapter(
     }
 
     override fun getItemCount(): Int = theaters.size
-
-    class TheaterViewHolder(
-        private val binding: ItemTheaterBinding,
-        private val movieContentId: Long,
-    ) :
-        RecyclerView.ViewHolder(binding.root) {
-        fun bind(theater: Theater) {
-            binding.theater = theater
-            binding.root.setOnClickListener {
-                val context = binding.root.context
-                Intent(context, MovieReservationActivity::class.java).apply {
-                    putExtra(MovieHomeKey.MOVIE_CONTENT_ID, movieContentId)
-                    putExtra(MovieHomeKey.THEATER_ID, theater.id)
-                    context.startActivity(this)
-                }
-            }
-        }
-    }
 }
